pyproject.toml
[project]
name = "manufacturing-analytics-spark"
version = "0.1.0"
dependencies = [
    "pyspark>=3.4.0",
    "delta-spark>=2.4.0",
    "openai>=1.0.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "pandas>=2.0.0",
    "google-generativeai>=0.8.6",
    "pytz>=2026.1.post1",
    "cerebras-cloud-sdk>=1.67.0",
    "tavily-python>=0.7.23",
    "groq>=1.2.0",
    "pytest>=9.0.3",
]
